Update Array helper APIs in ValuePropogationCommon for OffHeap
This commit is based on ed01b6dc36e7362986eda898463cc960135a82a5
Instead of modifying the code to support off-heap which as-is breaks
runs with arraylet environments, this commit adds separate paths for
off-heap while keeping the original paths as-is.

Code can be modified similar to the based-on commit to have gencon
runs using the array helper APIs while not affecting arraylet runs.

+++ compiler/optimizer/OMRTransformUtil.cpp
@@ -482,7 +482,7 @@ OMR::TransformUtil::generateArrayElementAddressTrees(TR::Compilation *comp, TR::
       }
    else
       {
-      totalOffsetNode = TR::Node::iconst(TR::Compiler->om.contiguousArrayHeaderSizeInBytes());
+      totalOffsetNode = TR::Node::iconst(static_cast<int32_t>(TR::Compiler->om.contiguousArrayHeaderSizeInBytes()));
       if (offsetNode)
          totalOffsetNode = TR::Node::create(TR::iadd, 2, offsetNode, totalOffsetNode);
       arrayAddressNode = TR::Node::create(TR::aiadd, 2, arrayNode, totalOffsetNode);
